Update `Elapsed` field type to `int64` in `scan_response.go`

+++ pkg/model/scan_response.go
@@ -1,22 +1,22 @@
 package model
 
 import (
-    "time"
+	"time"
 )
 
 type ResponseFileMeta struct {
-    FileId    string `json:"file_id"`
-    Name      string `json:"name"`
-    Sha256    string `json:"sha256"`
-    Size      int64  `json:"size"`
-    Verdict   string `json:"verdict"`
-    Signature string `json:"signature"`
-    Elapsed   int    `json:"elapsed"`
+	FileId    string `json:"file_id"`
+	Name      string `json:"name"`
+	Sha256    string `json:"sha256"`
+	Size      int64  `json:"size"`
+	Verdict   string `json:"verdict"`
+	Signature string `json:"signature"`
+	Elapsed   int64  `json:"elapsed"`
 }
 
 type ScanResponse struct {
-    JobId     string             `json:"job_id"`
-    Timestamp time.Time          `json:"timestamp"`
-    Elapsed   int                `json:"elapsed"`
-    Files     []ResponseFileMeta `json:"files"`
+	JobId     string             `json:"job_id"`
+	Timestamp time.Time          `json:"timestamp"`
+	Elapsed   int64              `json:"elapsed"`
+	Files     []ResponseFileMeta `json:"files"`
 }
